requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
oauthlib==3.2.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
tweepy==4.12.1
urllib3==1.26.14
